Promote Candidate B to main page, remove A/B test routes
- Replace old main page with Candidate B (Clean Pro theme)
- Delete /candidate_A, /candidate_B, /candidate_C routes
- Clean, professional UI now serves as primary experience
